LemParser.jjt: Commented out the 'description()' part of the bridge grammar. This is because an issue needs to be resolved with there being two 'descriptions' (model and bridge)
Bridge.java:
made this implement DescribedEntity so that it could be visited properly

BuilderPass1.java:
Added visit(LEMBridgeDeclaration) so that bridges can start being worked
into the models.

Bridges are now a part of the object model. They still need to be worked into
the GUI, and the issue with the two descriptions needs to be solved.