Added support for LEMVariableDeclarationList nodes. LEMVariableDeclarationList returns a LinkedList of VariableDeclaration actions.
Added support for LEMVariableDeclaration nodes. LEMVariableDeclaration
returns a VariableDeclaration Action.

Modified LEMProcedure so that the Procedure's ActionBlock is not created
as the result of a side-effect from the now-deleted visit() call.

Add support for LEMStatementList. LEMStatementList returns a LinkedList of
Actions.

Add support for LEMAction. LEMAction simply returns the result of visiting
its first child. This result should either be null (in the case of an
unimplemented action, in which case a warning is printed), or an instance
of Action to be returned.

Add support for LEMActionBlock. LEMActionBlocks visit their only two children:
a LEMVariableDeclarationList and a LEMStatementList. These two calls result in
two linked lists of Actions. These actions are added to a new ActionBlock, and
this ActionBlock is then returned.