BuilderPass2 now knows the class in which it is visiting nodes (if any). This is a bit of a dirty hack (but, I think, better than passing context through the visit() method).
Reformat code.

Add support for the LEMBehaviour node using the "new" method: visit the children
individually and add the resulting objects to the model (rather than relying
on children having side-effects)

Add a warning to indicate when children of LEMBehaviour do not return a
supported type.

Moved LEMTransitionDeclaration to the second pass, added a couple of helper methods