Object refcounting now working OK. Deletion of objects in stress test causes objects to eventually be collected, although there is still some continual memory usage probably due to the logger / gui.
Objects with state machines still have a few corner cases. Also, we should
not allow deletion of objects with state machines because of the possibility
of races versus the state machine threads.

We need the _last_ state machine that enters a deletion state to perform
the equivalent of a delete action. This requires a synchronized count of
state machines in the non deleted state.

State machines that enter quiescent states are going to keep the object
around indefinitely in the global pool. Presumably that is what the
action language modeller had intended.