Changed most of the gui classes and turned them into netBeans Form Beans, so that they are controllable and editable from the netBeans interface, then changed the position, layouts and minimum, maximum and preferred size of all the gui elements in order to give it a better shape.
Plus added signal injection to lem verifier, that means by clicking on the
listed Events under events subtree, of any model that has events defined on it,
the user gets the option of generating that signal to this object as a normal
signal, add that signal to this object as a self signal, add that signal to the
object as a delayed signal and add it to the object as a delayedSelfSignal. In
the case of the delayedSignals the user will receive an inputDialog in which
has to enter the amount of delay in miliseconds.